Blockchain service, implemented rate limiter and minor fixes
