+++ backend/.env.sample
@@ -7,6 +7,9 @@ SUPABASE_URL=https://yourprojectid.supabase.co
 SUPABASE_KEY="yourkey"
 CONTRACT_ADDRESS="0x2ff"
 CONTRACT_URL="yourapi"
+PRIVATE_KEY="yourprivatekey"
+EVENT_SERVER="http://localhost:3001"
+SOLIDITY_NODE="https://rinkeby.infura.io/v3/yourinfurakey"
 EMAIL_FROM="yourmail@service.com"
 EMAIL_CLIENT_ID='yourclientid'
 EMAIL_CLIENT_SECRET="yourclientsecret"
