+++ backend/.gitignore
@@ -9,3 +9,4 @@ build/
 .vscode/
 src/supabase/
 *.log
+./eng.traineddata
