+++ backend/.husky/_/pre-commit
@@ -1,2 +1,2 @@
 #!/usr/bin/env sh
-cd backend && npx lint-staged
+. "${0%/*}/h"
