+++ backend/package.json
@@ -9,7 +9,7 @@
         "start": "node ./build/index.js",
         "lint": "eslint . --ext .ts",
         "lint:fix": "eslint . --ext .ts --fix",
-        "prepare": "cd ../ && husky backend/.husky && cd -",
+        "prepare": "cd ../ && husky backend/.husky",
         "test": "jest"
     },
     "keywords": [],
@@ -19,13 +19,15 @@
         "*.{ts,yaml}": "prettier --write"
     },
     "dependencies": {
+        "@descope/node-sdk": "^1.6.6",
+        "@supabase/supabase-js": "^2.39.8",
         "bcrypt": "^5.1.1",
         "clamscan": "^2.2.1",
         "cookie-parser": "^1.4.6",
         "cors": "^2.8.5",
         "dotenv": "^16.4.5",
         "express": "^4.18.3",
-        "express-rate-limiter": "^1.3.1",
+        "express-rate-limit": "^7.2.0",
         "helmet": "^7.1.0",
         "looks-same": "^9.0.0",
         "morgan": "^1.10.0",
@@ -42,6 +44,7 @@
         "yamljs": "^0.3.0"
     },
     "devDependencies": {
+        "@daochild/tronweb-typescript": "^1.1.2",
         "@jest/globals": "^29.7.0",
         "@types/bcrypt": "^5.0.2",
         "@types/clamscan": "^2.0.8",
@@ -64,7 +67,6 @@
         "lint-staged": "^15.2.2",
         "nodemon": "^3.1.0",
         "supabase": "^1.148.6",
-        "tronweb-typings": "^1.0.1",
         "ts-jest": "^29.1.2",
         "ts-node": "^10.9.2",
         "typescript": "^5.4.2"
